large fix

+++ tasks/shvetsova_k_rad_sort_batch_merge/common/include/common.hpp
@@ -5,8 +5,8 @@
 
 #include "task/include/task.hpp"
 namespace shvetsova_k_rad_sort_batch_merge {
-using InType = std::vector<double>;
-using OutType = std::vector<double>;
+using InType = std::vector<int>;
+using OutType = std::vector<int>;
 using TestType = std::string;
 using BaseTask = ppc::task::Task<InType, OutType>;
 }  // namespace shvetsova_k_rad_sort_batch_merge
