+++ tasks/shvetsova_k_rad_sort_batch_merge/seq/include/ops_seq.hpp
@@ -16,7 +16,7 @@ class ShvetsovaKRadSortBatchMergeSEQ : public BaseTask {
   explicit ShvetsovaKRadSortBatchMergeSEQ(const InType &in);
 
  private:
-  std::vector<double> data_;
+  std::vector<int> data_;
 
   bool ValidationImpl() override;
   bool PreProcessingImpl() override;
@@ -24,10 +24,10 @@ class ShvetsovaKRadSortBatchMergeSEQ : public BaseTask {
   bool PostProcessingImpl() override;
 
   // доп функции //
-  static void RadixSort(std::vector<double> &vec);
-  static void BatcherOddEvenMergeSort(std::vector<double> &vec, int left, int right);
-  static void ExecuteBatcherStep(std::vector<double> &vec, int left, int n, int p, int k);
-  static void CompareAndSwap(std::vector<double> &vec, int i, int j);
+  static void RadixSort(std::vector<int> &vec);
+  static void BatcherOddEvenMergeSort(std::vector<int> &vec, int left, int right);
+  static void ExecuteBatcherStep(std::vector<int> &vec, int left, int n, int p, int k);
+  static void CompareAndSwap(std::vector<int> &vec, int i, int j);
 };
 
 }  // namespace shvetsova_k_rad_sort_batch_merge
